Define polygon properties before drawing anything
This was causing the very first 3D frame of the game to not show up the
text because the properties hadn't been set up. They would appear on the
second frame. However, the first frame was displayed for a second due to
it being the one being displayed during the initial fade in.